censorialchat was migrated to maven standard archetype, log4j was used for server loggingadd
