--- censorialchat/io/exception/UnableToReadException.java
@@ -1,7 +1,0 @@
-package JavaAutomationLab.censorialchat.io.exception;
-
-public class UnableToReadException extends Exception{
-    public UnableToReadException(String message) {
-        super(message);
-    }
-}
