--- censorialchat/io/exception/UnableToWriteException.java
@@ -1,7 +1,0 @@
-package JavaAutomationLab.censorialchat.io.exception;
-
-public class UnableToWriteException extends Exception{
-    public UnableToWriteException(String message) {
-        super(message);
-    }
-}
