--- censorialchat/io/interfaces/Packable.java
@@ -1,4 +1,0 @@
-package JavaAutomationLab.censorialchat.io.interfaces;
-
-public interface Packable {
-}
