--- censorialchat/io/interfaces/StreamReader.java
@@ -1,7 +1,0 @@
-package JavaAutomationLab.censorialchat.io.interfaces;
-
-import JavaAutomationLab.censorialchat.io.exception.UnableToReadException;
-
-public interface StreamReader {
-    Packable read() throws UnableToReadException, UnableToReadException;
-}
