--- config.properties
@@ -1,3 +1,0 @@
-hostname=127.0.0.1
-port=8000
-token=user
