+++ mavendemo/src/main/java/autolab/censorialchat/constant/C10Constant.java
@@ -1,0 +1,7 @@
+package autolab.censorialchat.constant;
+
+public class C10Constant {
+    public static final String HOSTNAME = "hostname";
+    public static final String PORT = "port";
+    public static final String TOKEN = "token";
+}
