+++ mavendemo/src/main/java/autolab/censorialchat/constant/TimeConstant.java
@@ -1,0 +1,5 @@
+package autolab.censorialchat.constant;
+
+public class TimeConstant {
+    public static final int TIME_TO_DELAY = 10000;
+}
