+++ mavendemo/src/main/java/autolab/censorialchat/io/exception/UnableToReadException.java
@@ -1,0 +1,7 @@
+package autolab.censorialchat.io.exception;
+
+public class UnableToReadException extends Exception{
+    public UnableToReadException(String message) {
+        super(message);
+    }
+}
