+++ mavendemo/src/main/java/autolab/censorialchat/io/impl/file/BufferedTextFileReader.java
@@ -1,0 +1,47 @@
+package autolab.censorialchat.io.impl.file;
+
+import autolab.censorialchat.io.base.BaseReader;
+import autolab.censorialchat.io.exception.UnableToCloseException;
+import autolab.censorialchat.io.interfaces.Reader;
+
+import java.io.*;
+
+public class BufferedTextFileReader extends BaseReader implements Reader {
+
+    public BufferedTextFileReader(File file) {
+        super(file);
+    }
+
+    public BufferedTextFileReader(String path) {
+        super(path);
+    }
+
+    public String read() throws UnableToCloseException {
+        BufferedReader objReader = null;
+        StringBuilder sb = new StringBuilder();
+        try {
+            String strCurrentLine;
+
+            objReader = new BufferedReader(new FileReader(this.path));
+
+            while ((strCurrentLine = objReader.readLine()) != null) {
+
+                sb.append(strCurrentLine);
+            }
+
+        } catch (IOException e) {
+            e.printStackTrace();
+
+        } finally {
+
+            try {
+                if (objReader != null)
+                    objReader.close();
+            } catch (IOException ex) {
+                ex.printStackTrace();
+                throw new UnableToCloseException("Unable to close");
+            }
+        }
+        return sb.toString();
+    }
+}
