+++ mavendemo/src/main/java/autolab/censorialchat/io/interfaces/Packable.java
@@ -1,0 +1,4 @@
+package autolab.censorialchat.io.interfaces;
+
+public interface Packable {
+}
