+++ mavendemo/src/main/java/autolab/censorialchat/io/interfaces/Reader.java
@@ -1,0 +1,8 @@
+package autolab.censorialchat.io.interfaces;
+
+import autolab.censorialchat.io.exception.UnableToCloseException;
+import autolab.censorialchat.io.exception.UnableToReadException;
+
+public interface Reader {
+    String read() throws UnableToReadException, UnableToCloseException;
+}
