+++ mavendemo/src/main/java/autolab/censorialchat/io/interfaces/StreamReader.java
@@ -1,0 +1,7 @@
+package autolab.censorialchat.io.interfaces;
+
+import autolab.censorialchat.io.exception.UnableToReadException;
+
+public interface StreamReader {
+    Packable read() throws UnableToReadException;
+}
