+++ mavendemo/src/main/java/autolab/censorialchat/io/interfaces/StreamWriter.java
@@ -1,0 +1,7 @@
+package autolab.censorialchat.io.interfaces;
+
+import autolab.censorialchat.io.exception.UnableToWriteException;
+
+public interface StreamWriter {
+    void write(String path, Packable pkg) throws UnableToWriteException;
+}
